Move DatabaseChangedEvent to separate file
