+++ extensions/ql-vscode/src/databases/local-databases/database-item-impl.ts
@@ -14,9 +14,9 @@ import { DatabaseItem, PersistedDatabaseItem } from "./database-item";
 import { isLikelyDatabaseRoot } from "../../helpers";
 import { stat } from "fs-extra";
 import { pathsEqual } from "../../pure/files";
-import { DatabaseChangedEvent, DatabaseEventKind } from "../local-databases";
 import { DatabaseContents } from "./database-contents";
 import { DatabaseResolver } from "./database-resolver";
+import { DatabaseChangedEvent, DatabaseEventKind } from "./database-events";
 
 export class DatabaseItemImpl implements DatabaseItem {
   private _error: Error | undefined = undefined;
